requirements.txt
Django==2.2.2
djangorestframework==3.9.4
python-dateutil==2.8.0
requests==2.22.0
